Implemented Functionality Changes
implemented the 1st functionality change of the issue:
I noticed that the stack ends with "May is on a call with Lea This call takes 20 minutes." and finally "Lea is on a call with May. This call takes 10 minutes." Notice how this is a cycle. It's supposed to end on Lea having no more calls. In other words, there's a misinterpretation in what this stack possibly represents. Please see the Recursion Lab Notes document for clarification.