Attempted to implement backtracking
I tried to implement the backtracking but I'm running into a few errors. I'm not entirely sure how to get it so that after the stack is done being called, the next button will then operate to be the trigger for the backtracking output. I also don't know how to get the base case of the backtracking button to be "Lea's call took 0 minutes." And lastly, the previous button will also not erase the first case, which I'm working to resolve. So far this is the issue I'm working on:
Implement backtracking. This means once the last person has no calls, then that means we can compute all of the previous people's "time until their call is finished".